[Fix]메인화면관련 리퀘스트 작성
