+++ src/main/java/com/multicampus/matchcode/service/khj/MypageService.java
@@ -1,6 +1,7 @@
 package com.multicampus.matchcode.service.khj;
 
 import com.multicampus.matchcode.model.entity.*;
+import com.multicampus.matchcode.model.request.khj.MemberAndPointRequest;
 import com.multicampus.matchcode.repository.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -31,50 +32,21 @@ public class MypageService {
 
 
     //메인화면 member 이름과 point를 불러오기 위한 service메소드
-    public MemberAndPointDTO getMemberAndPoint(long memberId) {
+    public MemberAndPointRequest getMemberAndPoint(long memberId) {
         Optional<MemberDTO> memberDTO = member.findById(memberId);
 
         if (memberDTO.isPresent()) {
             Optional<PointDTO> pointDTO = point.findByUserId(memberId);
 
             if (pointDTO.isPresent()) {
-                return new MemberAndPointDTO(memberDTO.get(), pointDTO.get());
+                return new MemberAndPointRequest(memberDTO.get(), pointDTO.get());
             } else {
                 // If point info is not found, you can pass an empty list or handle it based on your requirement
-                return new MemberAndPointDTO(memberDTO.get(), null);
+                return new MemberAndPointRequest(memberDTO.get(), null);
             }
         } else {
             return null;
         }
     }
-
-
-    //문제1) 그런데 '나의' 매치 기록을 가져와야 하는데...?
-
-    //매치 히스토리 화면 내 종목에 맞는 기록을 불러오기 위한 service메소드
-    public List<MatchDTO> getMatchesBySportsId(long sportsId) {
-        return matches.findBySportsId(sportsId);
-    }
-
-
-    public ResultDTO getResultByMatchId(long matchId) {
-        Optional<ResultDTO> odto = result.findByMatchIdAndUserId(matchId, (long)1);
-        if(odto.isPresent()){
-            return odto.get();
-        }
-        return null;
-    }
-    //결과는 매치id뿐만 아니라, 유저id까지 가져오면서 '내' 매치기록들만 읽어오도록
-
-
-    public MapDTO getMapByMatchId(long mapId) {
-        return map.findById(mapId);
-    }
-
-
-    public MatchMemberDTO getMatchMemberByMatchId(long matchid) {
-        return matchmember.findByMatchId(matchid);
-    }
-
 
 }
