+++ src/main/resources/templates/khj/mypage.html
@@ -18,9 +18,9 @@
     <h3>마이페이지</h3>
     (이미지가 들어갈 부분)
     <br>
-    이름 : <p style="display:inline;" th:text="${memberAndPoint.memberDTO.name}"></p>
+    이름 : <p style="display:inline;" th:text="${nameAndPoint.memberDTO.name}"></p>
     <br>
-    포인트 : <p style="display:inline;" th:text="${memberAndPoint.pointDTO.point}">p</p>
+    포인트 : <p style="display:inline;" th:text="${nameAndPoint.pointDTO.point}">p</p>
     <br>
     <button id="matchHistoryButton">매치 히스토리</button>
     <br>
